# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Genealogy Tree REST API",
        "title": "Genealogy Tree API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/trees/{tree_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение доступного пользователю дерева",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trees"
                ],
                "summary": "Получение дерева",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ULID дерева",
                        "name": "tree_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Дерево",
                        "schema": {
                            "$ref": "#/definitions/internal_features_trees_transport_http.TreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tree not found",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление дерева владельцем вместе с его участниками и связанными данными",
                "tags": [
                    "trees"
                ],
                "summary": "Удаление дерева",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ULID дерева",
                        "name": "tree_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Дерево удалено"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tree not found",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменение названия дерева владельцем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trees"
                ],
                "summary": "Изменение дерева",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ULID дерева",
                        "name": "tree_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTree тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_trees_transport_http.PatchTreeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Изменённое дерево",
                        "schema": {
                            "$ref": "#/definitions/internal_features_trees_transport_http.TreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tree not found",
                        "schema": {
                            "$ref": "#/definitions/genealogy-tree_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genealogy-tree_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
...
        "internal_features_trees_transport_http.PatchTreeRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Family Petrov Updated"
                }
            }
        },
        "internal_features_trees_transport_http.TreeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "01JQ2Q4K7Y8F6M2Z3N4P5R6S7T"
                },
                "name": {
                    "type": "string",
                    "example": "Family Petrov"
                },
                "owner_id": {
                    "type": "string",
                    "example": "01JQ2Q4K8Z8F6M2Z3N4P5R6S7U"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}